+++ docs/changelog.rst
@@ -60,6 +60,9 @@ New flavors:
 
 New fields for existing flavors:
 
+- Added CIN Number field in Morocco flavor (`gh-705 <https://github.com/django/django-localflavor/pull/507>`_).
+- Added get_states_of_brazil to br.utils to return a state or all available options with some settings
+  (`gh-510 <https://github.com/django/django-localflavor/pull/510>`_).
 - Added CIN Number field in Morocco flavor
   (`gh-705 <https://github.com/django/django-localflavor/pull/507>`_).